0910 cho - Order Update

+++ src/main/java/com/project/erpre/controller/CategoryController.java
@@ -28,7 +28,7 @@ public class CategoryController {
 
     //특정 카테고리
     @GetMapping("/{categoryNo}")
-    public Optional<Category> getCategoryById(@PathVariable Long categoryNo) {
+    public Optional<Category> getCategoryById(@PathVariable Integer categoryNo) {
         return categoryService.getCategoryById(categoryNo);
     }
 
@@ -40,7 +40,7 @@ public class CategoryController {
 
     //카테고리 수정
     @PutMapping("/{categoryNo}")
-    public Category updateCategory(@PathVariable Long categoryNo, @RequestBody Category category ) {
+    public Category updateCategory(@PathVariable Integer categoryNo, @RequestBody Category category ) {
         category.setCategoryNo(categoryNo);
         return null;
         //return categoryService.saveCategory(category);
@@ -48,7 +48,7 @@ public class CategoryController {
 
     //카테고리 삭제
     @DeleteMapping("/{categoryNo}")
-    public void deleteCategory(@PathVariable Long categoryNo) {
+    public void deleteCategory(@PathVariable Integer categoryNo) {
         categoryService.deleteCategory(categoryNo);
     }
 
@@ -59,13 +59,13 @@ public class CategoryController {
     }
 
     @GetMapping("/middle/{topCategoryId}")
-    public List<Category> getMiddleCategory(@PathVariable Long topCategoryId) {
+    public List<Category> getMiddleCategory(@PathVariable Integer topCategoryId) {
         return categoryService.getMiddleCategory(topCategoryId);
     }
 
     @GetMapping("/low/{middleCategoryId}/{topCategoryId}")
-    public List<Category> getLowCategory(@PathVariable Long topCategoryId,
-                                         @PathVariable Long middleCategoryId) {
+    public List<Category> getLowCategory(@PathVariable Integer topCategoryId,
+                                         @PathVariable Integer middleCategoryId) {
         return categoryService.getLowCategory(topCategoryId, middleCategoryId);
     }
