+++ src/main/java/com/project/erpre/model/Category.java
@@ -21,7 +21,7 @@ public class Category {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "category_no")
-    private Long categoryNo;
+    private Integer categoryNo;
 
     @Column(name = "category_level", nullable = false)
     private Integer categoryLevel = 1; // 기본값 설정
